Transliteration of radvd code into babeld code
In trying to get a grip on what the overall dependencies of
lua on uci, it seemed logical to try and get babeld equipped
with a gui interface finally.

This is not that in any way, merely a transliteration of the
existing radvd code. As I get a grip on how to use luci
properly, I'll get this working.