Skeleton for handling bulk updates and presets
It is painful to bulk network renumber, bulk change the SSIDs,
bulk change the domain name from any but the command line.

Each of those is a 1 line sed script. Around which I envision
a hundred lines of gui.

I also have a vague 'preset' concept, where you could configure
a box to be a home router, testbed box, interior gateway,
mesh node, etc. I would rather like to be able to do that,
scriptably - so I could bulk flash boxes, among other things.

This is a placeholder for that.